refactor: split unit panels, intro overlay, and region map lifecycle (#3878)
Extract build/flow logic into library parts for civilian and naval unit
panels, game-start intro overlay, and CtRegionMapGame onLoad/tap lifecycle
without changing public APIs or player-facing behavior.
