+++ app/lib/features/game/widgets/units/naval/naval_units_panel.dart
@@ -29,6 +29,7 @@ import '../shared/region_section_header.dart';
 import '../shared/region_labels.dart';
 
 part 'naval_units_panel_list.dart';
+part 'naval_units_panel_build.dart';
 part 'naval_units_panel_support_combine.dart';
 part 'naval_units_panel_support_home_transfer.dart';
 part 'naval_units_panel_support_dialogs.dart';
@@ -157,94 +158,5 @@ class _NavalUnitsPanelState extends BaseUnitsPanelState<NavalUnitsPanel> {
   }
 
   @override
-  Widget build(BuildContext context) {
-    final l10n = appL10n(context);
-    final tileScopeActive =
-        widget.tileScopeTileKey != null && widget.tileScopeTileKey!.isNotEmpty;
-    final tree = buildNavalTree(
-      widget.game,
-      widget.humanPlayerId,
-      widget.topology,
-      widget.draftOrders,
-      l10n,
-      tileMapByRegion: widget.tileMapByRegion,
-      topologyByRegion: widget.topologyByRegion,
-      locationScopeKeyFilter: widget.locationScopeKey,
-    );
-    final flat = flattenNavalTree(tree);
-    _visibleScopedFleetIds.clear();
-    _visibleScopedFleetIds.addAll(flat.map((row) => row.fleetId));
-    final hasAny = tree.any(
-      (group) => group.homeFleet != null || group.locations.isNotEmpty,
-    );
-    final canCombine = !widget.readOnly && _canCombineSelection(flat);
-    final readOnly = widget.readOnly;
-
-    // Header actions render as compact **primary** pills
-    // (`CtActionTextButton(primary: true)`) per SPEC/ui/naval-units-panel.md
-    // § Header actions and issue #3514 owner decisions #5 / #15. The optional
-    // tile-scope pill (and its 4px spacer) leads the shared select-all +
-    // Combine cluster assembled by `BaseUnitsPanelState.buildUnitsPanel`.
-    final panel = buildUnitsPanel(
-      title: tileScopeActive
-          ? l10n.naval_units_title_tile
-          : l10n.naval_units_title,
-      leadingActions: [
-        if (tileScopeActive)
-          CtActionTextButton(
-            primary: true,
-            enabled: widget.tileScopeTileKey!.isNotEmpty,
-            onPressed: () {
-              final key = widget.tileScopeTileKey!;
-              widget.bus.closePanelThenEmit(
-                OpenMapTileDetailEvent(tileKey: key),
-              );
-            },
-            label: l10n.civilian_units_tile,
-          ),
-        if (tileScopeActive && hasAny && flat.isNotEmpty)
-          const SizedBox(width: 4),
-      ],
-      trailingActions: [
-        CtActionTextButton(
-          primary: true,
-          onPressed: readOnly ? null : _openTrainDialog,
-          enabled: !readOnly,
-          label: l10n.common_train,
-        ),
-      ],
-      showCombineCluster: hasAny && flat.isNotEmpty && !readOnly,
-      selectableIds: _fleetSelectionIds(flat),
-      selectAllTooltip: l10n.naval_units_selectAllFleets,
-      deselectAllTooltip: l10n.naval_units_deselectAllFleets,
-      combineLabel: l10n.common_combine,
-      canCombine: canCombine,
-      onSelectAll: () => _onHeaderSelectAllTapped(flat),
-      onCombine: () => _performCombine(flat),
-      hasContent: hasAny,
-      listChildren: _navalListChildren(
-        tree: tree,
-        l10n: l10n,
-        readOnly: readOnly,
-      ),
-      emptyMessage: l10n.naval_units_empty,
-    );
-    if (kCtE2EEnabled) {
-      updateCtE2eNavalPanelSnapshotIfEnabled(
-        CtE2eNavalPanelSnapshot(
-          game: widget.game,
-          humanPlayerId: widget.humanPlayerId,
-          topology: widget.topology,
-          draftOrders: widget.draftOrders,
-          tileMapByRegion: widget.tileMapByRegion,
-          topologyByRegion: widget.topologyByRegion,
-          locationScopeKey: widget.locationScopeKey,
-          initialSelectedFleetId: widget.initialSelectedFleetId,
-          tileScopeTileKey: widget.tileScopeTileKey,
-        ),
-      );
-      return KeyedSubtree(key: kCtE2ENavalPanelRootKey, child: panel);
-    }
-    return panel;
-  }
+  Widget build(BuildContext context) => buildNavalUnitsPanel(context);
 }
